Added option to delete build files and mve ant-contrib option to build step configuration.

+++ src/main/java/hudson/plugins/antexec/AntExec.java
@@ -15,7 +15,7 @@
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * FITNESS FOR A PARTICULAR PURPOSE AND NON INFRINGEMENT. IN NO EVENT SHALL THE
  * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
@@ -33,9 +33,7 @@ import hudson.util.FormValidation;
 import hudson.tasks.Builder;
 import hudson.tasks.BuildStepDescriptor;
 import hudson.util.VariableResolver;
-import net.sf.json.JSONObject;
 import org.kohsuke.stapler.DataBoundConstructor;
-import org.kohsuke.stapler.StaplerRequest;
 import org.kohsuke.stapler.QueryParameter;
 import org.xml.sax.*;
 import org.xml.sax.helpers.XMLReaderFactory;
@@ -66,22 +64,26 @@ public class AntExec extends Builder {
     private final String scriptName;
     private final String properties;
     private final String antOpts;
+    private final Boolean keepBuildfile;
     private final Boolean verbose;
     private final Boolean emacs;
+    private final Boolean noAntcontrib;
     private final String antName;
 
     // Fields in config.groovy must match the parameter names in the "DataBoundConstructor"
     @SuppressWarnings("ALL")
     @DataBoundConstructor
-    public AntExec(String scriptSource, String extendedScriptSource, String scriptName, String properties, String antName, String antOpts, Boolean verbose, Boolean emacs) {
+    public AntExec(String scriptSource, String extendedScriptSource, String scriptName, String properties, String antName, String antOpts, Boolean keepBuildfile, Boolean verbose, Boolean emacs, Boolean noAntcontrib) {
         this.scriptSource = scriptSource;
         this.extendedScriptSource = extendedScriptSource;
         this.scriptName = scriptName;
         this.properties = properties;
+        this.keepBuildfile = keepBuildfile;
         this.antName = antName;
         this.antOpts = antOpts;
         this.verbose = verbose;
         this.emacs = emacs;
+        this.noAntcontrib = noAntcontrib;
     }
 
     /**
@@ -134,6 +136,15 @@ public class AntExec extends Builder {
     /**
      * Returns checkbox boolean from job configuration screen
      *
+     * @return Boolean keepBuildfile
+     */
+    public Boolean getKeepBuildfile() {
+        return keepBuildfile;
+    }
+
+    /**
+     * Returns checkbox boolean from job configuration screen
+     *
      * @return Boolean verbose
      */
     public Boolean getVerbose() {
@@ -143,10 +154,19 @@ public class AntExec extends Builder {
     /**
      * Returns checkbox boolean from job configuration screen
      *
-     * @return Boolean verbose
+     * @return Boolean emacs
      */
     public Boolean getEmacs() {
         return emacs;
+    }
+
+    /**
+     * Returns checkbox boolean from job configuration screen
+     *
+     * @return Boolean noAntcontrib
+     */
+    public Boolean getNoAntcontrib() {
+        return noAntcontrib;
     }
 
     /**
@@ -174,7 +194,7 @@ public class AntExec extends Builder {
         }
         EnvVars env = build.getEnvironment(listener);
         env.overrideAll(build.getBuildVariables());
- 
+
         Ant.AntInstallation ai = getAnt();
         if (ai == null) {
             args.add(launcher.isUnix() ? "ant" : "ant.bat");
@@ -188,7 +208,7 @@ public class AntExec extends Builder {
             args.add(exe);
         }
 
-        
+
         //Create Ant build.xml file
         FilePath buildFile = makeBuildFile(scriptName, scriptSourceResolved, extendedScriptSourceResolved, build);
 
@@ -200,10 +220,10 @@ public class AntExec extends Builder {
         @SuppressWarnings("unchecked") VariableResolver<String> vr = build.getBuildVariableResolver();
         @SuppressWarnings("unchecked") Set<String> sensitiveVars = build.getSensitiveBuildVariables();
         //noinspection unchecked
-              
+
         //Resolve the properties passed
         args.addKeyValuePairsFromPropertyString("-D", properties, vr, sensitiveVars);
-       
+
         if (ai != null)
             env.put("ANT_HOME", ai.getHome());
         if (antOpts != null && antOpts.length() > 0 && !antOpts.equals("")) {
@@ -211,7 +231,7 @@ public class AntExec extends Builder {
         }
 
         //Get and prepare ant-contrib.jar
-        if (getDescriptor().useAntContrib()) {
+        if (noAntcontrib != null && !noAntcontrib) {
             //TODO: Replace this with better methot
             if (verbose != null && verbose) listener.getLogger().println(Messages.AntExec_UseAntContribTasks());
             FilePath antContribJarOnMaster = new FilePath(Hudson.getInstance().getRootPath(), "plugins/antexec/META-INF/lib/ant-contrib.jar");
@@ -260,12 +280,11 @@ public class AntExec extends Builder {
                 r = launcher.launch().cmds(args).envs(env).stdout(aca).pwd(buildFile.getParent()).join();
             } finally {
                 aca.forceEol();
-                //After the ant script has been executed, we delete the build.xml. 
+                //After the ant script has been executed, we delete the build.xml.
                 //The plugin is a way to run an Ant Script from a small source code, we shoudn't keep the build.xml
-                boolean deleteResponse = buildFile.delete();
-                if(!deleteResponse)
-                {
-                    listener.getLogger().println("The temporary Ant Build Script coudn't be deleted");
+                if (keepBuildfile != null && !keepBuildfile) {
+                    boolean deleteResponse = buildFile.delete();
+                    if (!deleteResponse) listener.getLogger().println("The temporary Ant Build Script coudn't be deleted");
                 }
             }
             return r == 0;
@@ -294,7 +313,6 @@ public class AntExec extends Builder {
     @SuppressWarnings("UnusedDeclaration")
     @Extension
     public static final class DescriptorImpl extends BuildStepDescriptor<Builder> {
-        private boolean useAntContrib;
 
         @CopyOnWrite
         private volatile Ant.AntInstallation[] installations = new Ant.AntInstallation[0];
@@ -341,17 +359,6 @@ public class AntExec extends Builder {
 
         public String getDisplayName() {
             return Messages.AntExec_DisplayName();
-        }
-
-        @Override
-        public boolean configure(StaplerRequest req, JSONObject formData) throws FormException {
-            useAntContrib = formData.getBoolean("useAntContrib");
-            save();
-            return super.configure(req, formData);
-        }
-
-        public boolean useAntContrib() {
-            return useAntContrib;
         }
     }
